fix paths
