+++ docs/_modules/pygromos/hpc_queuing/job_scheduling/file_management.html
@@ -153,7 +153,7 @@
           <div role="main" class="document" itemscope="itemscope" itemtype="http://schema.org/Article">
            <div itemprop="articleBody">
             
-  <h1>Source code for pygromos.hpc_queuing.job_scheduling.file_management</h1><div class="highlight"><pre>
+  <h1>Source code for pygromos.simulations.hpc_queuing.job_scheduling.file_management</h1><div class="highlight"><pre>
 <span></span><span class="sd">&quot;&quot;&quot;</span>
 <span class="sd">    This module is doing all the post simulation file juggeling needed for gromos. CURRENTLY OLD DON&quot;T USE</span>
 <span class="sd">&quot;&quot;&quot;</span>
