+++ docs/_modules/pygromos/hpc_queuing/job_scheduling/schedulers/scheduler_functions.html
@@ -153,7 +153,7 @@
           <div role="main" class="document" itemscope="itemscope" itemtype="http://schema.org/Article">
            <div itemprop="articleBody">
             
-  <h1>Source code for pygromos.hpc_queuing.job_scheduling.schedulers.scheduler_functions</h1><div class="highlight"><pre>
+  <h1>Source code for pygromos.simulations.hpc_queuing.job_scheduling.schedulers.scheduler_functions</h1><div class="highlight"><pre>
 <span></span><span class="kn">import</span> <span class="nn">glob</span>
 <span class="kn">import</span> <span class="nn">os</span>
