+++ docs/_modules/pygromos/hpc_queuing/job_scheduling/schedulers/simulation_scheduler.html
@@ -153,7 +153,7 @@
           <div role="main" class="document" itemscope="itemscope" itemtype="http://schema.org/Article">
            <div itemprop="articleBody">
             
-  <h1>Source code for pygromos.hpc_queuing.job_scheduling.schedulers.simulation_scheduler</h1><div class="highlight"><pre>
+  <h1>Source code for pygromos.simulations.hpc_queuing.job_scheduling.schedulers.simulation_scheduler</h1><div class="highlight"><pre>
 <span></span><span class="sd">&quot;&quot;&quot;</span>
 <span class="sd">Wrapper for Long simulations -  similar to repex_EDS_long_production_run.</span>
 <span class="sd">This script schedules Simulations ons euler into the queue.</span>
