+++ docs/_modules/pygromos/hpc_queuing/job_scheduling/workers/simulation_workers/clean_up_simulation_files.html
@@ -153,7 +153,7 @@
           <div role="main" class="document" itemscope="itemscope" itemtype="http://schema.org/Article">
            <div itemprop="articleBody">
             
-  <h1>Source code for pygromos.hpc_queuing.job_scheduling.workers.simulation_workers.clean_up_simulation_files</h1><div class="highlight"><pre>
+  <h1>Source code for pygromos.simulations.hpc_queuing.job_scheduling.workers.simulation_workers.clean_up_simulation_files</h1><div class="highlight"><pre>
 <span></span><span class="ch">#!/usr/bin/env python3</span>
 
 <span class="sd">&quot;&quot;&quot; After simulation cleanup</span>
