+++ docs/_modules/pygromos/hpc_queuing/job_scheduling/workers/simulation_workers/simulation_run_worker.html
@@ -153,7 +153,7 @@
           <div role="main" class="document" itemscope="itemscope" itemtype="http://schema.org/Article">
            <div itemprop="articleBody">
             
-  <h1>Source code for pygromos.hpc_queuing.job_scheduling.workers.simulation_workers.simulation_run_worker</h1><div class="highlight"><pre>
+  <h1>Source code for pygromos.simulations.hpc_queuing.job_scheduling.workers.simulation_workers.simulation_run_worker</h1><div class="highlight"><pre>
 <span></span><span class="ch">#!/usr/bin/env python3</span>
 <span class="kn">import</span> <span class="nn">os</span>
 <span class="kn">import</span> <span class="nn">sys</span>
