+++ docs/_modules/pygromos/hpc_queuing/submission_systems.html
@@ -153,7 +153,7 @@
           <div role="main" class="document" itemscope="itemscope" itemtype="http://schema.org/Article">
            <div itemprop="articleBody">
             
-  <h1>Source code for pygromos.hpc_queuing.submission_systems</h1><div class="highlight"><pre>
+  <h1>Source code for pygromos.simulations.hpc_queuing.submission_systems</h1><div class="highlight"><pre>
 <span></span><span class="kn">import</span> <span class="nn">socket</span>
 
 <span class="kn">from</span> <span class="nn">pygromos.hpc_queuing.submission_systems.dummy</span> <span class="kn">import</span> <span class="n">DUMMY</span>
