+++ docs/_modules/pygromos/hpc_queuing/submission_systems/dummy.html
@@ -155,7 +155,7 @@
           <div role="main" class="document" itemscope="itemscope" itemtype="http://schema.org/Article">
            <div itemprop="articleBody">
             
-  <h1>Source code for pygromos.hpc_queuing.submission_systems.dummy</h1><div class="highlight"><pre>
+  <h1>Source code for pygromos.simulations.hpc_queuing.submission_systems.dummy</h1><div class="highlight"><pre>
 <span></span><span class="kn">import</span> <span class="nn">warnings</span>
 <span class="kn">from</span> <span class="nn">typing</span> <span class="kn">import</span> <span class="n">Union</span><span class="p">,</span> <span class="n">List</span>
