+++ docs/_modules/pygromos/hpc_queuing/submission_systems/local.html
@@ -155,7 +155,7 @@
           <div role="main" class="document" itemscope="itemscope" itemtype="http://schema.org/Article">
            <div itemprop="articleBody">
             
-  <h1>Source code for pygromos.hpc_queuing.submission_systems.local</h1><div class="highlight"><pre>
+  <h1>Source code for pygromos.simulations.hpc_queuing.submission_systems.local</h1><div class="highlight"><pre>
 <span></span><span class="kn">import</span> <span class="nn">os</span>
 <span class="kn">import</span> <span class="nn">warnings</span>
 <span class="kn">import</span> <span class="nn">pandas</span> <span class="k">as</span> <span class="nn">pd</span>
