+++ docs/_modules/pygromos/hpc_queuing/submission_systems/lsf.html
@@ -155,7 +155,7 @@
           <div role="main" class="document" itemscope="itemscope" itemtype="http://schema.org/Article">
            <div itemprop="articleBody">
             
-  <h1>Source code for pygromos.hpc_queuing.submission_systems.lsf</h1><div class="highlight"><pre>
+  <h1>Source code for pygromos.simulations.hpc_queuing.submission_systems.lsf</h1><div class="highlight"><pre>
 <span></span><span class="kn">import</span> <span class="nn">os</span>
 <span class="kn">import</span> <span class="nn">re</span>
 <span class="kn">import</span> <span class="nn">time</span>
