+++ docs/_sources/Examples/example_gromos_simulation.ipynb.txt
@@ -176,7 +176,7 @@
     }
    ],
    "source": [
-    "from pygromos.hpc_queuing.submission_systems.local import LOCAL as subSystem\n",
+    "from pygromos.simulations.hpc_queuing.submission_systems.local import LOCAL as subSystem\n",
     "from pygromos.simulations.modules.preset_simulation_modules import emin\n",
     "step_name  = \"a_vacuum_emin\"\n",
     "\n",
