+++ docs/_sources/Examples/example_submission_systems.ipynb.txt
@@ -17,9 +17,9 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "from pygromos.hpc_queuing.submission_systems import local # this executes your code in your local session.\n",
-    "from pygromos.hpc_queuing.submission_systems import lsf # this module can be used to submit to the lsf-queue (e.g. on euler)\n",
-    "from pygromos.hpc_queuing.submission_systems import dummy # this is a dummy system, that only prints the commands"
+    "from pygromos.simulations.hpc_queuing.submission_systems import local # this executes your code in your local session.\n",
+    "from pygromos.simulations.hpc_queuing.submission_systems import lsf # this module can be used to submit to the lsf-queue (e.g. on euler)\n",
+    "from pygromos.simulations.hpc_queuing.submission_systems import dummy # this is a dummy system, that only prints the commands"
    ]
   },
   {
