+++ docs/_sources/_source/pygromos.hpc_queuing.job_scheduling.workers.analysis_workers.analysis_tools.rst.txt
@@ -4,7 +4,7 @@ pygromos.hpc\_queuing.job\_scheduling.workers.analysis\_workers.analysis\_tools 
 Module contents
 ---------------
 
-.. automodule:: pygromos.hpc_queuing.job_scheduling.workers.analysis_workers.analysis_tools
+.. automodule:: pygromos.simulations.hpc_queuing.job_scheduling.workers.analysis_workers.analysis_tools
    :members:
    :undoc-members:
    :show-inheritance:
