+++ docs/_sources/_source/pygromos.hpc_queuing.job_scheduling.workers.analysis_workers.rst.txt
@@ -7,7 +7,7 @@ Subpackages
 .. toctree::
    :maxdepth: 4
 
-   pygromos.hpc_queuing.job_scheduling.workers.analysis_workers.analysis_tools
+   pygromos.simulations.hpc_queuing.job_scheduling.workers.analysis_workers.analysis_tools
 
 Submodules
 ----------
@@ -15,7 +15,7 @@ Submodules
 pygromos.hpc\_queuing.job\_scheduling.workers.analysis\_workers.simulation\_analysis module
 -------------------------------------------------------------------------------------------
 
-.. automodule:: pygromos.hpc_queuing.job_scheduling.workers.analysis_workers.simulation_analysis
+.. automodule:: pygromos.simulations.hpc_queuing.job_scheduling.workers.analysis_workers.simulation_analysis
    :members:
    :undoc-members:
    :show-inheritance:
@@ -23,7 +23,7 @@ pygromos.hpc\_queuing.job\_scheduling.workers.analysis\_workers.simulation\_anal
 Module contents
 ---------------
 
-.. automodule:: pygromos.hpc_queuing.job_scheduling.workers.analysis_workers
+.. automodule:: pygromos.simulations.hpc_queuing.job_scheduling.workers.analysis_workers
    :members:
    :undoc-members:
    :show-inheritance:
