+++ docs/_sources/_source/pygromos.hpc_queuing.job_scheduling.workers.rst.txt
@@ -7,13 +7,13 @@ Subpackages
 .. toctree::
    :maxdepth: 4
 
-   pygromos.hpc_queuing.job_scheduling.workers.analysis_workers
-   pygromos.hpc_queuing.job_scheduling.workers.simulation_workers
+   pygromos.simulations.hpc_queuing.job_scheduling.workers.analysis_workers
+   pygromos.simulations.hpc_queuing.job_scheduling.workers.simulation_workers
 
 Module contents
 ---------------
 
-.. automodule:: pygromos.hpc_queuing.job_scheduling.workers
+.. automodule:: pygromos.simulations.hpc_queuing.job_scheduling.workers
    :members:
    :undoc-members:
    :show-inheritance:
