+++ docs/_sources/_source/pygromos.hpc_queuing.job_scheduling.workers.simulation_workers.rst.txt
@@ -7,7 +7,7 @@ Submodules
 pygromos.hpc\_queuing.job\_scheduling.workers.simulation\_workers.clean\_up\_simulation\_files module
 -----------------------------------------------------------------------------------------------------
 
-.. automodule:: pygromos.hpc_queuing.job_scheduling.workers.simulation_workers.clean_up_simulation_files
+.. automodule:: pygromos.simulations.hpc_queuing.job_scheduling.workers.simulation_workers.clean_up_simulation_files
    :members:
    :undoc-members:
    :show-inheritance:
@@ -15,7 +15,7 @@ pygromos.hpc\_queuing.job\_scheduling.workers.simulation\_workers.clean\_up\_sim
 pygromos.hpc\_queuing.job\_scheduling.workers.simulation\_workers.simulation\_run\_worker module
 ------------------------------------------------------------------------------------------------
 
-.. automodule:: pygromos.hpc_queuing.job_scheduling.workers.simulation_workers.simulation_run_worker
+.. automodule:: pygromos.simulations.hpc_queuing.job_scheduling.workers.simulation_workers.simulation_run_worker
    :members:
    :undoc-members:
    :show-inheritance:
@@ -23,7 +23,7 @@ pygromos.hpc\_queuing.job\_scheduling.workers.simulation\_workers.simulation\_ru
 Module contents
 ---------------
 
-.. automodule:: pygromos.hpc_queuing.job_scheduling.workers.simulation_workers
+.. automodule:: pygromos.simulations.hpc_queuing.job_scheduling.workers.simulation_workers
    :members:
    :undoc-members:
    :show-inheritance:
