+++ docs/_sources/_source/pygromos.hpc_queuing.rst.txt
@@ -7,13 +7,13 @@ Subpackages
 .. toctree::
    :maxdepth: 4
 
-   pygromos.hpc_queuing.job_scheduling
-   pygromos.hpc_queuing.submission_systems
+   pygromos.simulations.hpc_queuing.job_scheduling
+   pygromos.simulations.hpc_queuing.submission_systems
 
 Module contents
 ---------------
 
-.. automodule:: pygromos.hpc_queuing
+.. automodule:: pygromos.simulations.hpc_queuing
    :members:
    :undoc-members:
    :show-inheritance:
