+++ docs/_sources/_source/pygromos.hpc_queuing.submission_systems.rst.txt
@@ -7,7 +7,7 @@ Submodules
 pygromos.hpc\_queuing.submission\_systems.dummy module
 ------------------------------------------------------
 
-.. automodule:: pygromos.hpc_queuing.submission_systems.dummy
+.. automodule:: pygromos.simulations.hpc_queuing.submission_systems.dummy
    :members:
    :undoc-members:
    :show-inheritance:
@@ -15,7 +15,7 @@ pygromos.hpc\_queuing.submission\_systems.dummy module
 pygromos.hpc\_queuing.submission\_systems.local module
 ------------------------------------------------------
 
-.. automodule:: pygromos.hpc_queuing.submission_systems.local
+.. automodule:: pygromos.simulations.hpc_queuing.submission_systems.local
    :members:
    :undoc-members:
    :show-inheritance:
@@ -23,7 +23,7 @@ pygromos.hpc\_queuing.submission\_systems.local module
 pygromos.hpc\_queuing.submission\_systems.lsf module
 ----------------------------------------------------
 
-.. automodule:: pygromos.hpc_queuing.submission_systems.lsf
+.. automodule:: pygromos.simulations.hpc_queuing.submission_systems.lsf
    :members:
    :undoc-members:
    :show-inheritance:
@@ -31,7 +31,7 @@ pygromos.hpc\_queuing.submission\_systems.lsf module
 Module contents
 ---------------
 
-.. automodule:: pygromos.hpc_queuing.submission_systems
+.. automodule:: pygromos.simulations.hpc_queuing.submission_systems
    :members:
    :undoc-members:
    :show-inheritance:
