+++ pygromos/simulations/hpc_queuing/submission_systems/local.py
@@ -3,7 +3,7 @@ import warnings
 import pandas as pd
 from typing import Union, List
 
-from pygromos.hpc_queuing.submission_systems._submission_system import _SubmissionSystem
+from pygromos.simulations.hpc_queuing.submission_systems._submission_system import _SubmissionSystem
 from pygromos.utils import bash
