+++ pygromos/simulations/hpc_queuing/submission_systems/lsf.py
@@ -1,11 +1,9 @@
 import os
-import re
-import time
 from datetime import datetime
 import pandas as pd
 from typing import Union, List
 
-from pygromos.hpc_queuing.submission_systems._submission_system import _SubmissionSystem
+from pygromos.simulations.hpc_queuing.submission_systems._submission_system import _SubmissionSystem
 from pygromos.utils import bash
